refactor cognitive tests
