+++ cognitive/src/test/scala/com/microsoft/azure/synapse/ml/cognitive/CognitiveServicesCommon.scala
