+++ cognitive/src/test/scala/com/microsoft/azure/synapse/ml/cognitive/face/FaceAPI.scala
@@ -1,7 +1,7 @@
 // Copyright (C) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in project root for information.
 
-package com.microsoft.azure.synapse.ml.cognitive.split1
+package com.microsoft.azure.synapse.ml.cognitive.face
 
 import com.microsoft.azure.synapse.ml.cognitive._
 import com.microsoft.azure.synapse.ml.core.env.StreamUtilities.using
@@ -78,7 +78,7 @@ object FaceUtils extends CognitiveKey {
   }
 }
 
-import com.microsoft.azure.synapse.ml.cognitive.split1.FaceUtils._
+import com.microsoft.azure.synapse.ml.cognitive.face.FaceUtils._
 
 object FaceListProtocol {
   implicit val PfiEnc = jsonFormat2(PersistedFaceInfo.apply)
